organization: separate files for boundaryconditionmanager
